pyproject.toml
[tool.poetry]
name = "dspy-wordpress"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.13.3"
dspy-ai = "^2.3.6"
weaviate-client = "^4.5.1"
python-dotenv = "^1.0.1"
rag4p = "^0.1.3"
rockset = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
